+++ framework/domain/src/main/java/is/codion/framework/domain/entity/EmptyEntity.java
@@ -1,0 +1,58 @@
+/*
+ * This file is part of Codion.
+ *
+ * Codion is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Codion is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Codion.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * Copyright (c) 2025, Björn Darri Sigurðsson.
+ */
+package is.codion.framework.domain.entity;
+
+import java.io.IOException;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
+import java.io.Serial;
+
+import static is.codion.framework.domain.entity.EntitySerializer.serializerForDomain;
+import static java.util.Objects.requireNonNull;
+
+final class EmptyEntity extends ImmutableEntity {
+
+	@Serial
+	private static final long serialVersionUID = 1;
+
+	private String toString;
+
+	EmptyEntity(EntityDefinition definition, String toString) {
+		super(new DefaultEntity(definition));
+		this.toString = requireNonNull(toString);
+	}
+
+	@Override
+	public String toString() {
+		return toString;
+	}
+
+	@Serial
+	private void writeObject(ObjectOutputStream stream) throws IOException {
+		stream.writeObject(definition.type().domainType().name());
+		EntitySerializer.serialize(this, stream);
+		stream.writeObject(toString);
+	}
+
+	@Serial
+	private void readObject(ObjectInputStream stream) throws IOException, ClassNotFoundException {
+		serializerForDomain((String) stream.readObject()).deserialize(this, stream);
+		toString = (String) stream.readObject();
+	}
+}